@@ -79,8 +79,8 @@
     "promise-polyfill": "8.1.0",
     "prop-types": "^15.6.1",
     "query-string": "^6.1.0",
-    "react": "^16.6.3",
-    "react-dom": "^16.6.3",
+    "react": "^16.8.2",
+    "react-dom": "^16.8.2",
     "react-immutable-proptypes": "^2.1.0",
     "react-infinite-scroller": "^1.1.4",
     "react-modal": "^3.4.4",
