@@ -1,29 +1,48 @@
-import React, { PureComponent } from 'react';
+import React from 'react';
 import InsideBanner from './InsideBanner';
 
-class About extends PureComponent {
-  render() {
-    return (
-      <main className="about">
-        <InsideBanner title="About Us" />
-        <section className="container">
-          <div className="spacer">
-            <div className="row">
-              <div className="col-lg-8  col-lg-offset-2">
-                <img src="images/about.jpg" className="img-responsive thumbnail" alt="realestate" />
-                <h3>Business Background</h3>
-                <p>Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry&#39;s standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.</p>
-                <h3>Company Profile</h3>
-                <p>Contrary to popular belief, Lorem Ipsum is not simply random text. It has roots in a piece of classical Latin literature from 45 BC, making it over 2000 years old. Richard McClintock, a Latin professor at Hampden-Sydney College in Virginia, looked up one of the more obscure Latin words, consectetur, from a Lorem Ipsum passage, and going through the cites of the word in classical literature, discovered the undoubtable source. Lorem Ipsum comes from sections 1.10.32 and 1.10.33 of &quot;de Finibus Bonorum et Malorum&quot; (The Extremes of Good and Evil) by Cicero, written in 45 BC. This book is a treatise on the theory of ethics, very popular during the Renaissance. The first line of Lorem Ipsum, &quot;Lorem ipsum dolor sit amet..&quot;, comes from a line in section 1.10.32.</p>
-                <p>The standard chunk of Lorem Ipsum used since the 1500s is reproduced below for those interested. Sections 1.10.32 and 1.10.33 from &quot;de Finibus Bonorum et Malorum&quot; by Cicero are also reproduced in their exact original form, accompanied by English versions from the 1914 translation by H. Rackham.</p>
-              </div>
-            </div>
+const About = () => (
+  <main className="about">
+    <InsideBanner title="About Us" />
+    <section className="container">
+      <div className="spacer">
+        <div className="row">
+          <div className="col-lg-8  col-lg-offset-2">
+            <img src="images/about.jpg" className="img-responsive thumbnail" alt="realestate" />
+            <h3>Business Background</h3>
+            <p>
+              Lorem Ipsum is simply dummy text of the printing and typesetting industry.
+              Lorem Ipsum has been the industry&#39;s standard dummy text ever since the 1500s,
+              when an unknown printer took a galley of type and scrambled it to make a type specimen
+              book. It has survived not only five centuries, but also the leap into electronic
+              typesetting, remaining essentially unchanged. It was popularised in the 1960s with
+              the release of Letraset sheets containing Lorem Ipsum passages, and more recently with
+              desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.
+            </p>
+            <h3>Company Profile</h3>
+            <p>
+              Contrary to popular belief, Lorem Ipsum is not simply random text. It has roots in
+              a piece of classical Latin literature from 45 BC, making it over 2000 years old.
+              Richard McClintock, a Latin professor at Hampden-Sydney College in Virginia, looked
+              up one of the more obscure Latin words, consectetur, from a Lorem Ipsum passage,
+              and going through the cites of the word in classical literature, discovered the
+              undoubtable source. Lorem Ipsum comes from sections 1.10.32 and 1.10.33 of
+              &quot;de Finibus Bonorum et Malorum&quot; (The Extremes of Good and Evil) by Cicero,
+              written in 45 BC. This book is a treatise on the theory of ethics, very popular during
+              the Renaissance. The first line of Lorem Ipsum, &quot;Lorem ipsum dolor sit amet.
+              &quot;, comes from a line in section 1.10.32.
+            </p>
+            <p>
+              The standard chunk of Lorem Ipsum used since the 1500s is reproduced below for those
+              interested. Sections 1.10.32 and 1.10.33 from &quot;de Finibus Bonorum et
+              Malorum&quot; by Cicero are also reproduced in their exact original form,
+              accompanied by English versions from the 1914 translation by H. Rackham.
+            </p>
           </div>
-        </section>
-      </main>
-    );
-  }
-}
-
+        </div>
+      </div>
+    </section>
+  </main>
+);
 
 export default About;
