@@ -1,4 +1,4 @@
-import React, { PureComponent } from 'react';
+import React, { useEffect } from 'react';
 import { PropTypes } from 'prop-types';
 import { connect } from 'react-redux';
 import { updateLoginModalIsOpen } from '../actions/layoutActions';
@@ -6,32 +6,33 @@ import InsideBanner from '../components/InsideBanner';
 import UserRegisterForm from '../components/UserRegisterForm';
 import { userRegisterValidate } from '../actions/userActions';
 
-class RegisterContainer extends PureComponent {
-  componentDidMount() {
-    if (this.props.loginModalOpen) {
-      this.props.updateLoginModalIsOpen(false);
+const RegisterContainer = ({
+  loginModalOpen,
+  registerValidateAction,
+  updateLoginModalAction,
+}) => {
+  useEffect(() => {
+    if (loginModalOpen) {
+      updateLoginModalAction(false)
     }
-  }
+  });
 
-  render() {
-    const { registerValidate } = this.props;
-    return (
-      <main className="register">
-        <InsideBanner title="Register" />
-        <section className="container">
-          <div className="row register">
-            <UserRegisterForm onValidate={registerValidate} />
-          </div>
-        </section>
-      </main>
-    );
-  }
-}
+  return (
+    <main className="register">
+      <InsideBanner title="Register" />
+      <section className="container">
+        <div className="row register">
+          <UserRegisterForm onValidate={registerValidateAction} />
+        </div>
+      </section>
+    </main>
+  );
+};
 
 RegisterContainer.propTypes = {
   loginModalOpen: PropTypes.bool.isRequired,
-  updateLoginModalIsOpen: PropTypes.func.isRequired,
-  registerValidate: PropTypes.func.isRequired,
+  updateLoginModalAction: PropTypes.func.isRequired,
+  registerValidateAction: PropTypes.func.isRequired,
 };
 
 const mapStateToProps = state => ({
@@ -39,8 +40,8 @@ const mapStateToProps = state => ({
 });
 
 const functions = {
-  updateLoginModalIsOpen,
-  registerValidate: userRegisterValidate,
+  updateLoginModalAction: updateLoginModalIsOpen,
+  registerValidateAction: userRegisterValidate,
 };
 
 export default connect(mapStateToProps, functions)(RegisterContainer);
